chore: extract common funcs
